+++ smk-parent/smks-api/src/main/java/ca/bc/gov/databc/smks/model/LayerBase.java
@@ -1,0 +1,206 @@
+package ca.bc.gov.databc.smks.model;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.annotation.JsonSubTypes;
+import com.fasterxml.jackson.annotation.JsonSubTypes.Type;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+
+import ca.bc.gov.databc.smks.model.layer.EsriDynamic;
+import ca.bc.gov.databc.smks.model.layer.Folder;
+import ca.bc.gov.databc.smks.model.layer.Group;
+import ca.bc.gov.databc.smks.model.layer.Wms;
+import ca.bc.gov.databc.smks.model.layer.Kml;
+import ca.bc.gov.databc.smks.model.layer.Geojson;
+
+@JsonTypeInfo(
+	use = JsonTypeInfo.Id.NAME,
+	include = JsonTypeInfo.As.PROPERTY,
+	property = "type"
+)
+@JsonSubTypes( {
+	@Type( name = "esri-dynamic", value = EsriDynamic.class ),
+	@Type( name = "folder",       value = Folder.class ),
+	@Type( name = "group",        value = Group.class ),
+	@Type( name = "wms",          value = Wms.class ),
+	@Type( name = "kml",          value = Kml.class ),
+	@Type( name = "geojson",      value = Geojson.class )
+} )
+@JsonInclude(Include.NON_NULL)
+public class LayerBase implements Cloneable
+{
+	// private Integer id;
+	private String id;
+	// private String label;
+	private String title;
+	private Boolean isVisible;
+	private String attribution;
+	private String metadataUrl;
+	// private List<Attribute> attributes;
+	// private String serviceUrl; // also used as path for KML and JSON
+	// private String format;
+	private Double opacity;
+	// private Boolean isTransparent;
+	private Double minScale;
+	private Double maxScale;
+
+	public LayerBase()
+	{
+	}
+
+	public String getId()
+	{
+		return id;
+	}
+
+	public void setId(String id)
+	{
+		this.id = id;
+	}
+
+	public String getTitle()
+	{
+		return title;
+	}
+
+	public void setTitle(String title)
+	{
+		this.title = title;
+	}
+
+	public boolean getIsVisible()
+	{
+		return isVisible;
+	}
+
+	public void setIsVisible(boolean isVisible)
+	{
+		this.isVisible = isVisible;
+	}
+
+	public String getAttribution()
+	{
+		return attribution;
+	}
+
+	public void setAttribution(String attribution)
+	{
+		this.attribution = attribution;
+	}
+
+	public String getMetadataUrl()
+	{
+		return metadataUrl;
+	}
+
+	public void setMetadataUrl(String metadataUrl)
+	{
+		this.metadataUrl = metadataUrl;
+	}
+
+	// public String getServiceUrl()
+	// {
+	// 	return serviceUrl;
+	// }
+
+	// public void setServiceUrl(String serviceUrl)
+	// {
+	// 	this.serviceUrl = serviceUrl;
+	// }
+
+	// public String getFormat()
+	// {
+	// 	return format;
+	// }
+
+	// public void setFormat(String format)
+	// {
+	// 	this.format = format;
+	// }
+
+	public Double getOpacity()
+	{
+		return opacity;
+	}
+
+	public void setOpacity(Double opacity)
+	{
+		this.opacity = opacity;
+	}
+
+	// public Boolean getIsTransparent()
+	// {
+	// 	return isTransparent;
+	// }
+
+	// public void setIsTransparent(Boolean isTransparent)
+	// {
+	// 	this.isTransparent = isTransparent;
+	// }
+
+	public Double getMinScale()
+	{
+		return minScale;
+	}
+
+	public void setMinScale(Double minScale)
+	{
+		this.minScale = minScale;
+	}
+
+	public Double getMaxScale()
+	{
+		return maxScale;
+	}
+
+	public void setMaxScale(Double maxScale)
+	{
+		this.maxScale = maxScale;
+	}
+
+	@Override
+	public String toString()
+	{
+		return this.getTitle();
+	}
+
+	// public List<Attribute> getAttributes()
+	// {
+	// 	if(attributes == null) attributes = new ArrayList<Attribute>();
+	// 	return attributes;
+	// }
+
+	// public void setAttributes(List<Attribute> attributes)
+	// {
+	// 	this.attributes = attributes;
+	// }
+
+	public LayerBase clone()
+	{
+		LayerBase clone = new LayerBase();
+
+		clone.setId(id);
+		clone.setTitle(title);
+		clone.setAttribution(attribution);
+		clone.setMetadataUrl(metadataUrl);
+		// clone.setFormat(format);
+		// clone.setIsTransparent(isTransparent);
+		clone.setIsVisible(isVisible);
+		// clone.setLabel(label);
+		clone.setMaxScale(maxScale);
+		clone.setMinScale(minScale);
+		clone.setOpacity(opacity);
+		// clone.setServiceUrl(serviceUrl);
+
+		// for(Attribute a : attributes)
+		// {
+		// 	clone.getAttributes().add(a.clone());
+		// }
+
+		return clone;
+	}
+
+}
