+++ smk-parent/smks-api/src/main/java/ca/bc/gov/databc/smks/model/layer/Feature.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.databc.smks.model;
+package ca.bc.gov.databc.smks.model.layer;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,28 +6,32 @@ import java.util.List;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
+import ca.bc.gov.databc.smks.model.Layer;
+
 @JsonInclude(Include.NON_NULL)
-public class FeatureLayer extends Layer
+public class Feature extends Layer
 {
 	private List<String> layers;
-	
-	public FeatureLayer() { }
-	
-	public List<String> getLayers() 
+
+public Feature() { }
+
+	public List<String> getLayers()
 	{
 		if(layers == null) layers = new ArrayList<String>();
 		return layers;
 	}
 
-	public void setLayers(List<String> layers) 
+	public void setLayers(List<String> layers)
 	{
 		this.layers = layers;
 	}
-	
-	public FeatureLayer clone()
+
+	public Feature clone()
 	{
-		FeatureLayer clone = new FeatureLayer();
-		
+		Feature clone = new Feature();
+
+
+
 		clone.setAttribution(getAttribution());
 		clone.setFormat(getFormat());
 		clone.setId(getId());
@@ -38,17 +42,17 @@ public class FeatureLayer extends Layer
 		clone.setMinScale(getMinScale());
 		clone.setOpacity(getOpacity());
 		clone.setServiceUrl(getServiceUrl());
-		
+
 		for(String layer : layers)
 		{
 			clone.getLayers().add(layer);
 		}
-		
+
 		for(Attribute a : getAttributes())
 		{
 			clone.getAttributes().add(a.clone());
 		}
-		
+
 		return clone;
 	}
 }
