+++ smk-parent/smks-api/src/main/java/ca/bc/gov/databc/smks/model/layer/Geojson.java
